change: exchange header and menubar
