+++ src/components/Header.tsx
@@ -1,39 +1,100 @@
-import { Box, Flex, Input, useColorMode } from "@chakra-ui/react";
-import { useAtom } from "jotai";
+import { useState } from "react";
 
-import { familyKeywordAtom } from "~/utils/jotai";
+import {
+  Box,
+  Flex,
+  IconButton,
+  Input,
+  InputGroup,
+  InputRightElement,
+  Menu,
+  MenuButton,
+  MenuItem,
+  MenuList,
+  Slider,
+  SliderFilledTrack,
+  SliderThumb,
+  SliderTrack,
+} from "@chakra-ui/react";
+import { CaretDown } from "@phosphor-icons/react";
+import { useAtom, useSetAtom } from "jotai";
+
+import { fontSizeAtom, previewStringAtom } from "~/utils/jotai";
 
 const Header: React.FC = () => {
-  const { colorMode, toggleColorMode } = useColorMode();
-  const [familyKeyword, setFamilyKeyword] = useAtom(familyKeywordAtom);
+  const setFontSize = useSetAtom(fontSizeAtom);
+  const [localFontSize, setLocalFontSize] = useState<number>(32);
+  const [previewString, setPreviewString] = useAtom(previewStringAtom);
+
+  const onChange = (e: number) => {
+    setLocalFontSize(e);
+    setFontSize(e);
+  };
+
+  const sampleTexts = [
+    "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz",
+    "0123456789 ¿ ? ¡ ! & @ ‘ ’ “ ” « » % * ^ # $ £ € ¢ / ( ) [ ] { } . , ® ©",
+    "The quick brown fox jumps over the lazy dog.",
+    "あのイーハトーヴォのすきとおった風、夏でも底に冷たさをもつ青いそら、うつくしい森で飾られたモリーオ市、郊外のぎらぎらひかる草の波。",
+  ];
 
   return (
-    <Box
-      as="header"
+    <Flex
       bg="gray.900"
-      sx={{
-        position: "sticky",
-        top: 0,
-        left: 0,
-        right: 0,
-        zIndex: 10,
-        p: "0.5rem",
-      }}
+      position="sticky"
+      top={0}
+      left={0}
+      right={0}
+      zIndex={15}
+      p="0.5rem"
+      gap="1rem"
+      h="4rem"
     >
-      <Flex
-        h="3rem"
-        alignItems={"center"}
-        justifyContent={"space-between"}
-        gap="0.5rem"
-      >
-        <Input
-          placeholder="フォントを検索"
-          value={familyKeyword}
-          onChange={(e) => setFamilyKeyword(e.target.value)}
-          variant="outline"
-        />
+      <Flex w="50%">
+        <Box w="4rem" m="auto">
+          {localFontSize}px
+        </Box>
+        <Slider
+          aria-label="slider-ex-1"
+          defaultValue={localFontSize}
+          onChange={onChange}
+          min={10}
+          max={100}
+          focusThumbOnChange={false}
+        >
+          <SliderTrack>
+            <SliderFilledTrack />
+          </SliderTrack>
+          <SliderThumb />
+        </Slider>
       </Flex>
-    </Box>
+      <Flex w="50%">
+        <InputGroup>
+          <Input
+            placeholder="プレビューする文字"
+            value={previewString}
+            onChange={(e) => setPreviewString(e.target.value)}
+            variant="outline"
+          />
+          <InputRightElement>
+            <Menu>
+              <MenuButton
+                as={IconButton}
+                icon={<CaretDown />}
+                variant="ghost"
+              />
+              <MenuList>
+                {sampleTexts.map((text) => (
+                  <MenuItem key={text} onClick={() => setPreviewString(text)}>
+                    {text}
+                  </MenuItem>
+                ))}
+              </MenuList>
+            </Menu>
+          </InputRightElement>
+        </InputGroup>
+      </Flex>
+    </Flex>
   );
 };
